Moved proxy to /api/proxy
