+++ src/app.module.ts
@@ -21,6 +21,7 @@ import { ToolsModule } from './providers/tools/module/tools.module.js';
 import { CustomThrottlerGuard } from './shared/throttler.guard.js';
 import { MappingsModule } from './providers/anime/mappings/module/mappings.module.js';
 import { AnilibriaModule } from './providers/anime/anilibria/module/anilibria.module.js';
+import { ProxyModule } from './providers/proxy/module/proxy.module.js';
 
 @Module({
   imports: [
@@ -38,6 +39,7 @@ import { AnilibriaModule } from './providers/anime/anilibria/module/anilibria.mo
     ToolsModule,
     ConsoleModule,
     MappingsModule,
+    ProxyModule,
     ConfigModule.forRoot({
       envFilePath: '.env',
       isGlobal: true,
